+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - `is_test` Added for Daemon Builders, when set to `true` will use temporary file for state
 - Chain configs now can be edited from a networks config file. It will read `~/.cw-orchestrator/networks.json`, see example `networks.json.example`
+- New package `cw-orch-neutron-test-tube`
 
 ### Breaking